backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
psycopg==3.1.19
gunicorn==22.0.0

django-extensions==3.2.3
Pillow==10.4.0
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
pytest-mock==3.14.0
dj-database-url==2.1.0
django-filter==24.2